Fix players being killed when jumping out of lane
There was a known issue of players being killed (plugin intervention)
when jumping out of their lane. This was due to the fact that the plugin
attempted to determine if the player was standing on solid ground by
checking the fall-speed. Adding a manual check for whether or not the
player is standing on a non-air block fixes this issue.