Fix bug with enderman block pickup on worlds where mob spawn is restricted
SPECULATION: Minecraft seems to allow endermen to pickup/interact with
blocks at some strange time right as they spawn. This means that if an
enderman spawns, attempts to immediately pickup a block, and then the
spawn is cancelled, the EntityBlockChangeEvent won't have an entity to
refer to in its callback handler.

The block, however, is guaranteed to exist, so just changing the code to
get the World from the block, not the entity.