Implement some of the final bits of BEDSONLY code.
Player.getBedSpawnLocation() now fulfills its contract to certify that
(1) the player has set a bed, (2) the location still contains a bed, and
(3) the bed is not obstructed or otherwise unavailable. Therefore, it is
sufficient to use this for all of our needs.

CraftBukkit fixed this bug in CB 1.4.7-R0.2 b2622/b2623